+++ R/print.treats.R
@@ -65,8 +65,10 @@ print.treats <- function(x, all = FALSE, ...) {
             cat(paste0(" ---- treats object with ", length(x), " replicates ---- \n"))
             cat(paste0("You can access individual replicates by extracting them, e.g. using x[[1]]\n"))
             x <- x[[1]]
+            replicated <- TRUE
         } else {
             cat(" ---- treats object ---- \n")
+            replicated <- FALSE
         }
 
         ## Print the modifiers and/or events
@@ -85,19 +87,25 @@ print.treats <- function(x, all = FALSE, ...) {
             if(!is.null(x$events)) {
                 lapply(x$events, print.events.info)
             }
+            # cat("\n")
+        }
+
+        if(!replicated) {
+            ## Print the tree
+            cat("Simulated phylogenetic tree (x$tree):\n")
+            cat(print(x$tree))
             cat("\n")
+
+            ## Print the traits
+            if(!is.null(x$data)) {
+                cat("Simulated trait data (x$data)")
+            }
         }
 
-        ## Print the tree
-        cat("Simulated phylogenetic tree (x$tree):\n")
-        cat(print(x$tree))
-        cat("\n")
-
-        ## Print the traits
-        if(!is.null(x$data)) {
-            cat("Simulated trait data (x$data)")
-        }
         if(!is.null(x$traits)) {
+            if(replicated) {
+                cat("Simulated trait data")
+            }
             cat(":\n")
             print.traits.info(x$traits$main)
             if(!is.null(x$traits$background)) {