+++ tests/testthat/test-treats.R
@@ -109,7 +109,7 @@ test_that("paper example works", {
 
     ## multiple prints
     output <- capture_output(print.treats(sim_rand_extinction))
-    expect_equal(output, " ---- treats object with 10 replicates ---- \nYou can access individual replicates by extracting them, e.g. using x[[1]]\nBirth death process with events:\nspeciation: 0.352.\nextinction: 0.288.\nEvent targeting \"taxa\" to be triggered 1 time.\nThe condition function is: time.condition\nThe modification function is: random.extinction\n\n\nSimulated phylogenetic tree (x$tree):\n\nPhylogenetic tree with 62 tips and 71 internal nodes.\n\nTip labels:\n  t1, t2, t3, t4, t5, t6, ...\nNode labels:\n  n1, n2, n3, n4, n5, n6, ...\n\nRooted; includes branch lengths.\n\nSimulated trait data (x$data):\n2 traits for 1 process (A:2) with one starting value (0).")
+    expect_equal(output, " ---- treats object with 10 replicates ---- \nYou can access individual replicates by extracting them, e.g. using x[[1]]\nBirth death process with events:\nspeciation: 0.352.\nextinction: 0.288.\nEvent targeting \"taxa\" to be triggered 1 time.\nThe condition function is: time.condition\nThe modification function is: random.extinction\n\nSimulated trait data:\n2 traits for 1 process (A:2) with one starting value (0).")
 
 
 
@@ -127,7 +127,7 @@ test_that("paper example works", {
     expect_equal(length(sim_trait_extinction), 10)
 
     output <- capture_output(print.treats(sim_trait_extinction))
-    expect_equal(output, " ---- treats object with 10 replicates ---- \nYou can access individual replicates by extracting them, e.g. using x[[1]]\nBirth death process with events:\nspeciation: 0.352.\nextinction: 0.288.\nEvent targeting \"taxa\" to be triggered 1 time.\nThe condition function is: time.condition\nThe modification function is: trait.extinction\n\n\nSimulated phylogenetic tree (x$tree):\n\nPhylogenetic tree with 101 tips and 109 internal nodes.\n\nTip labels:\n  t1, t2, t3, t4, t5, t6, ...\nNode labels:\n  n1, n2, n3, n4, n5, n6, ...\n\nRooted; includes branch lengths.\n\nSimulated trait data (x$data):\n2 traits for 1 process (A:2) with one starting value (0).")
+    expect_equal(output, " ---- treats object with 10 replicates ---- \nYou can access individual replicates by extracting them, e.g. using x[[1]]\nBirth death process with events:\nspeciation: 0.352.\nextinction: 0.288.\nEvent targeting \"taxa\" to be triggered 1 time.\nThe condition function is: time.condition\nThe modification function is: trait.extinction\n\nSimulated trait data:\n2 traits for 1 process (A:2) with one starting value (0).")
 
 })